8: adjust description of `git submodule update`
The current description of the behaviour of `git submodule update`
in fact corresponds to the behaviour of `git submodule update --remote`.

Update the description to reflect the actual behaviour of `git submodule
update`.

Instead of using `git submodule deinit` to remove the already
initialized submodule, create a new clone instead, which is a more
realistic example.

Also, introduce the "superproject" terminology.

Move the description of `git submodule update --remote` to the end of
the "Discussion" section.

Closes MikeMcQuaid/GitInPractice#18